fix: Issue with previous apprentices layout
Previously, there was an issue using `space-between` to lay out our grid
when there were two or three previous apprentice group sections on the
last row. When using `space-between` the layout would evenly distribute
the remaining row space between two and three items, which would cause
the sections to be vertically misaligned. This fixes that issue by
adding support for `grid` (when supported).